build.sh: add --no-backup-if-mismatch to `patch` command [UX]
The --no-backup-if-mismatch option tells `patch` never to generate a
index.html.orig backup file, even when the patch fails to apply
cleanly. In the past, I have accidentally included this file in my
Piglet releases and I don't want that. Besides, I don't need backups
because I use git for that.

The code highlighting patch often fails to apply cleanly
(or just fails entirely) as a result of me adding/removing text from
the manual. The proper fix would be to use pandoc-emphasize-code [1]
rather than a handwritten HTML patch, but pandoc-emphasize-code
doesn't work with the same version of pandoc that required by
pandoc-crossref.